feat: move more of AuthContext to wallet package
