+++ src/packages/wallet/hooks/useChainId/index.tsx
@@ -1,0 +1,3 @@
+import { store } from '../../store';
+
+export const useChainId = () => store.use.chainId();
