+++ src/packages/wallet/index.ts
@@ -1,9 +1,12 @@
 export * from './connectors';
 export * from './types';
 export * from './Web3Wrapper';
+export * from './AuthModal';
 export * from './chains';
 export * from './utilities/canAddTokenToWallet';
 export * from './utilities/addTokenToWallet';
 export * from './hooks/useProvider';
 export * from './hooks/useSigner';
 export * from './hooks/useAccountAddress';
+export * from './hooks/useSwitchChain';
+export * from './hooks/useChainId';
